+++ test/setup-test-env-angular.ts
@@ -1,6 +1,0 @@
-import "@testing-library/jest-dom";
-import "@angular/compiler";
-import "@analogjs/vitest-angular/setup-snapshots";
-import { setupTestBed } from "@analogjs/vitest-angular/setup-testbed";
-
-setupTestBed();